Sharded setup methods. Separate unit tests for GOAP setup.
