+++ src/engine/core/objects/ai/setup/state/animstate_planner.test.ts
@@ -1,0 +1,63 @@
+import { describe, it } from "@jest/globals";
+
+import { setupStalkerStatePlanner } from "@/engine/core/objects/ai";
+import { EStateActionId, EStateEvaluatorId } from "@/engine/core/objects/animation/state_types";
+import { ActionAnimationStart, ActionAnimationStop } from "@/engine/core/objects/state/animation";
+import { ActionAnimstateStart, ActionAnimstateStop } from "@/engine/core/objects/state/animstate";
+import { StalkerStateManager } from "@/engine/core/objects/state/StalkerStateManager";
+import { ActionPlanner, ClientObject } from "@/engine/lib/types";
+import { checkPlannerAction } from "@/fixtures/engine";
+import { mockClientGameObject } from "@/fixtures/xray";
+
+describe("animation_planner util", () => {
+  it("should correctly setup state planner animstate actions", () => {
+    const object: ClientObject = mockClientGameObject();
+    const stateManager: StalkerStateManager = new StalkerStateManager(object);
+    const planner: ActionPlanner = stateManager.planner;
+
+    setupStalkerStatePlanner(planner, stateManager);
+
+    checkPlannerAction(
+      planner.action(EStateActionId.ANIMSTATE_START),
+      ActionAnimstateStart,
+      [
+        [EStateEvaluatorId.LOCKED, false],
+        [EStateEvaluatorId.LOCKED_EXTERNAL, false],
+        [EStateEvaluatorId.ANIMSTATE, false],
+        [EStateEvaluatorId.SMARTCOVER, true],
+        [EStateEvaluatorId.ANIMATION_NONE_NOW, true],
+        [EStateEvaluatorId.DIRECTION, true],
+        [EStateEvaluatorId.MENTAL, true],
+        [EStateEvaluatorId.WEAPON, true],
+        [EStateEvaluatorId.MOVEMENT, true],
+        [EStateEvaluatorId.ANIMSTATE_PLAY_NOW, false],
+      ],
+      [[EStateEvaluatorId.ANIMSTATE, true]]
+    );
+
+    checkPlannerAction(
+      planner.action(EStateActionId.ANIMSTATE_STOP),
+      ActionAnimstateStop,
+      [
+        [EStateEvaluatorId.LOCKED, false],
+        [EStateEvaluatorId.LOCKED_EXTERNAL, false],
+        [EStateEvaluatorId.ANIMATION_LOCKED, false],
+        [EStateEvaluatorId.ANIMSTATE_LOCKED, false],
+        [EStateEvaluatorId.ANIMSTATE_IDLE_NOW, false],
+        [EStateEvaluatorId.ANIMATION_PLAY_NOW, false],
+      ],
+      [
+        [EStateEvaluatorId.ANIMSTATE, true],
+        [EStateEvaluatorId.ANIMSTATE_PLAY_NOW, false],
+        [EStateEvaluatorId.ANIMSTATE_IDLE_NOW, true],
+      ]
+    );
+
+    checkPlannerAction(
+      planner.action(EStateActionId.LOCKED_ANIMSTATE),
+      "ActionStateLockedAnimstate",
+      [[EStateEvaluatorId.ANIMSTATE_LOCKED, true]],
+      [[EStateEvaluatorId.ANIMSTATE_LOCKED, false]]
+    );
+  });
+});
