+++ src/engine/core/objects/ai/setup/state/direction_planner.test.ts
@@ -1,0 +1,53 @@
+import { describe, it } from "@jest/globals";
+
+import { setupStalkerStatePlanner } from "@/engine/core/objects/ai";
+import { EStateActionId, EStateEvaluatorId } from "@/engine/core/objects/animation/state_types";
+import { ActionDirectionSearch, ActionDirectionTurn } from "@/engine/core/objects/state/direction";
+import { StalkerStateManager } from "@/engine/core/objects/state/StalkerStateManager";
+import { ActionPlanner, ClientObject } from "@/engine/lib/types";
+import { checkPlannerAction } from "@/fixtures/engine";
+import { mockClientGameObject } from "@/fixtures/xray";
+
+describe("direction_planner util", () => {
+  it("should correctly setup state planner planner actions", () => {
+    const object: ClientObject = mockClientGameObject();
+    const stateManager: StalkerStateManager = new StalkerStateManager(object);
+    const planner: ActionPlanner = stateManager.planner;
+
+    setupStalkerStatePlanner(planner, stateManager);
+
+    checkPlannerAction(
+      planner.action(EStateActionId.DIRECTION_TURN),
+      ActionDirectionTurn,
+      [
+        [EStateEvaluatorId.ANIMSTATE_LOCKED, false],
+        [EStateEvaluatorId.ANIMATION_LOCKED, false],
+        [EStateEvaluatorId.LOCKED_EXTERNAL, false],
+        [EStateEvaluatorId.DIRECTION, false],
+        [EStateEvaluatorId.DIRECTION_SEARCH, false],
+        [EStateEvaluatorId.WEAPON, true],
+        [EStateEvaluatorId.ANIMATION_NONE_NOW, true],
+        [EStateEvaluatorId.ANIMSTATE_IDLE_NOW, true],
+        [EStateEvaluatorId.MOVEMENT, true],
+      ],
+      [[EStateEvaluatorId.DIRECTION, true]]
+    );
+
+    checkPlannerAction(
+      planner.action(EStateActionId.DIRECTION_SEARCH),
+      ActionDirectionSearch,
+      [
+        [EStateEvaluatorId.ANIMSTATE_LOCKED, false],
+        [EStateEvaluatorId.ANIMATION_LOCKED, false],
+        [EStateEvaluatorId.LOCKED_EXTERNAL, false],
+        [EStateEvaluatorId.DIRECTION, false],
+        [EStateEvaluatorId.DIRECTION_SEARCH, true],
+        [EStateEvaluatorId.WEAPON, true],
+        [EStateEvaluatorId.ANIMATION_NONE_NOW, true],
+        [EStateEvaluatorId.ANIMSTATE_IDLE_NOW, true],
+        [EStateEvaluatorId.MOVEMENT, true],
+      ],
+      [[EStateEvaluatorId.DIRECTION, true]]
+    );
+  });
+});
