+++ src/engine/core/objects/ai/setup/state/locked_planner.ts
@@ -6,18 +6,19 @@ import { ActionStateLocked } from "@/engine/core/objects/state/state";
 import { ActionPlanner } from "@/engine/lib/types";
 
 /**
+ * Setup GOAP logics related to locked state changes of stalkers logics.
  *
- * @param planner
- * @param stateManager
+ * @param planner - action planner to configure
+ * @param stateManager - target object state manager
  */
 export function setupStalkerLockedStatePlanner(planner: ActionPlanner, stateManager: StalkerStateManager): void {
-  const lockedAction = new ActionStateLocked(stateManager, "ActionStateLocked");
+  const lockedAction: ActionStateLocked = new ActionStateLocked(stateManager, "ActionStateLocked");
 
   lockedAction.add_precondition(new world_property(EStateEvaluatorId.LOCKED, true));
   lockedAction.add_effect(new world_property(EStateEvaluatorId.LOCKED, false));
   planner.add_action(EStateActionId.LOCKED, lockedAction);
 
-  const lockedExternalAction = new ActionStateLocked(stateManager, "ActionStateLockedExternal");
+  const lockedExternalAction: ActionStateLocked = new ActionStateLocked(stateManager, "ActionStateLockedExternal");
 
   lockedExternalAction.add_precondition(new world_property(EStateEvaluatorId.LOCKED_EXTERNAL, true));
   lockedExternalAction.add_effect(new world_property(EStateEvaluatorId.LOCKED_EXTERNAL, false));
